requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.11.7
    # via black
aiosignal==1.3.1
    # via aiohttp
asgiref==3.8.1
    # via
    #   django
    #   django-allauth
    #   django-htmx
attrs==24.2.0
    # via aiohttp
black[d]==24.10.0
    # via -r requirements.in
build==1.2.2.post1
    # via pip-tools
click==8.1.7
    # via
    #   black
    #   pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
django==5.1.3
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-debug-toolbar
    #   django-htmx
django-allauth==65.2.0
    # via -r requirements.in
django-cleanup==9.0.0
    # via -r requirements.in
django-debug-toolbar==4.2.0
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-htmx==1.21.0
    # via -r requirements.in
django-widget-tweaks==1.5.0
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via yarl
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
packaging==24.2
    # via
    #   black
    #   build
pathspec==0.12.1
    # via black
pillow==11.0.0
    # via -r requirements.in
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.6
    # via black
propcache==0.2.0
    # via
    #   aiohttp
    #   yarl
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
sqlparse==0.5.2
    # via
    #   django
    #   django-debug-toolbar
tzdata==2024.2
    # via django
wheel==0.45.1
    # via pip-tools
yarl==1.18.0
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
